falta integração com api
