+++ src/assets/styles/Card.scss
