+++ src/assets/styles/CardBig.scss
