+++ src/assets/styles/MainBigCard.scss
@@ -2,7 +2,6 @@
     width: 100%;
     height: 151px;
     background-image: linear-gradient(180deg, #ee4c77 10%, #0d0430 300%);
-    
     display: flex;
     flex-direction: row;
     justify-content: space-between;
@@ -12,4 +11,14 @@
     align-self: center;
     height: 40px;
     margin-left: 20px;
+}
+
+.bigCard{
+    display: flex;
+}
+
+.site{ 
+    background-color: #ebe9d7;
+    height: 97.6vh;
+
 }
