+++ src/containers/App.js
@@ -1,5 +1,6 @@
 import React from "react";
-import Routes from "./routes";
+import Routes from "../routes/routes";
+
 
 const App = () => <Routes />;
