+++ src/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
+import App from './containers/App';
 
 ReactDOM.render(<App />, document.getElementById('root'));
