+++ src/routes/routes.js
@@ -1,11 +1,11 @@
 import React from "react";
 import { BrowserRouter, Route, Switch, Redirect } from "react-router-dom";
-import App from './App'
-import Login from './services/Login'
-import Main from './Main'
-import MainBigCard from './MainBigCard'
+import App from '../containers/App'
+import Login from '../containers/Login'
+import Main from '../containers/Main'
+import MainBigCard from '../containers/MainBigCard'
 
-import { isAuthenticated } from "./services/auth";
+import { isAuthenticated } from "../services/auth";
 
 const PrivateRoute = ({ component: Component, ...rest }) => (
   <Route
